Internal: move utils to utils folder
