+++ modules/atomic-widgets/controls/types/image-control.php
@@ -2,7 +2,7 @@
 namespace Elementor\Modules\AtomicWidgets\Controls\Types;
 
 use Elementor\Modules\AtomicWidgets\Base\Atomic_Control_Base;
-use Elementor\Modules\AtomicWidgets\Image\Image_Sizes;
+use Elementor\Modules\AtomicWidgets\Utils\Image_Sizes;
 
 if ( ! defined( 'ABSPATH' ) ) {
 	exit; // Exit if accessed directly.
