+++ modules/atomic-widgets/controls/types/svg-control.php
@@ -3,7 +3,7 @@
 namespace Elementor\Modules\AtomicWidgets\Controls\Types;
 
 use Elementor\Modules\AtomicWidgets\Base\Atomic_Control_Base;
-use Elementor\Modules\AtomicWidgets\Image_Sizes;
+use Elementor\Modules\AtomicWidgets\Utils\Image_Sizes;
 
 if ( ! defined( 'ABSPATH' ) ) {
 	exit; // Exit if accessed directly.
