+++ modules/atomic-widgets/prop-types/image-prop-type.php
@@ -2,7 +2,7 @@
 
 namespace Elementor\Modules\AtomicWidgets\PropTypes;
 
-use Elementor\Modules\AtomicWidgets\Image\Image_Sizes;
+use Elementor\Modules\AtomicWidgets\Utils\Image_Sizes;
 use Elementor\Modules\AtomicWidgets\PropTypes\Base\Object_Prop_Type;
 use Elementor\Modules\AtomicWidgets\PropTypes\Primitives\String_Prop_Type;
